+++ modules/apps/commerce/commerce-order-content-web/src/main/java/com/liferay/commerce/order/content/web/internal/portlet/action/UploadCSVFileEntryMVCActionCommand.java
@@ -6,14 +6,35 @@
 package com.liferay.commerce.order.content.web.internal.portlet.action;
 
 import com.liferay.commerce.constants.CommercePortletKeys;
-import com.liferay.commerce.order.content.web.internal.upload.CSVUploadFileEntryHandler;
-import com.liferay.commerce.order.content.web.internal.upload.CSVUploadResponseHandler;
+import com.liferay.document.library.kernel.exception.FileExtensionException;
+import com.liferay.document.library.kernel.util.DLValidator;
+import com.liferay.item.selector.ItemSelectorUploadResponseHandler;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.portal.kernel.json.JSONUtil;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.model.Company;
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
+import com.liferay.portal.kernel.repository.model.FileEntry;
+import com.liferay.portal.kernel.servlet.ServletResponseConstants;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.upload.UploadPortletRequest;
+import com.liferay.portal.kernel.util.File;
+import com.liferay.portal.kernel.util.TempFileEntryUtil;
+import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.upload.UniqueFileNameProvider;
+import com.liferay.upload.UploadFileEntryHandler;
 import com.liferay.upload.UploadHandler;
+import com.liferay.upload.UploadResponseHandler;
+
+import java.io.IOException;
+import java.io.InputStream;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
+import javax.portlet.PortletRequest;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -30,6 +51,119 @@ import org.osgi.service.component.annotations.Reference;
 )
 public class UploadCSVFileEntryMVCActionCommand extends BaseMVCActionCommand {
 
+	public class CSVUploadFileEntryHandler implements UploadFileEntryHandler {
+
+		@Override
+		public FileEntry upload(UploadPortletRequest uploadPortletRequest)
+			throws IOException, PortalException {
+
+			ThemeDisplay themeDisplay =
+				(ThemeDisplay)uploadPortletRequest.getAttribute(
+					WebKeys.THEME_DISPLAY);
+
+			String fileName = uploadPortletRequest.getFileName(_PARAMETER_NAME);
+
+			_dlValidator.validateFileSize(
+				themeDisplay.getScopeGroupId(), fileName,
+				uploadPortletRequest.getContentType(_PARAMETER_NAME),
+				uploadPortletRequest.getSize(_PARAMETER_NAME));
+
+			String extension = _file.getExtension(fileName);
+
+			if (!extension.equals("csv")) {
+				throw new FileExtensionException(
+					"Invalid extension for file name " + fileName);
+			}
+
+			try (InputStream inputStream = uploadPortletRequest.getFileAsStream(
+					_PARAMETER_NAME)) {
+
+				return _addFileEntry(
+					fileName,
+					uploadPortletRequest.getContentType(_PARAMETER_NAME),
+					inputStream,
+					(ThemeDisplay)uploadPortletRequest.getAttribute(
+						WebKeys.THEME_DISPLAY));
+			}
+		}
+
+		private FileEntry _addFileEntry(
+				String fileName, String contentType, InputStream inputStream,
+				ThemeDisplay themeDisplay)
+			throws PortalException {
+
+			String uniqueFileName = _uniqueFileNameProvider.provide(
+				fileName, curFileName -> _exists(curFileName, themeDisplay));
+
+			Company company = themeDisplay.getCompany();
+
+			return TempFileEntryUtil.addTempFileEntry(
+				company.getGroupId(), themeDisplay.getUserId(),
+				_TEMP_FOLDER_NAME, uniqueFileName, inputStream, contentType);
+		}
+
+		private boolean _exists(String fileName, ThemeDisplay themeDisplay) {
+			try {
+				FileEntry tempFileEntry = TempFileEntryUtil.getTempFileEntry(
+					themeDisplay.getScopeGroupId(), themeDisplay.getUserId(),
+					_TEMP_FOLDER_NAME, fileName);
+
+				if (tempFileEntry != null) {
+					return true;
+				}
+
+				return false;
+			}
+			catch (PortalException portalException) {
+				if (_log.isDebugEnabled()) {
+					_log.debug(portalException);
+				}
+
+				return false;
+			}
+		}
+
+	}
+
+	public class CSVUploadResponseHandler implements UploadResponseHandler {
+
+		@Override
+		public JSONObject onFailure(
+				PortletRequest portletRequest, PortalException portalException)
+			throws PortalException {
+
+			JSONObject jsonObject =
+				_itemSelectorUploadResponseHandler.onFailure(
+					portletRequest, portalException);
+
+			if (portalException instanceof FileExtensionException) {
+				jsonObject.put(
+					"error",
+					JSONUtil.put(
+						"errorType",
+						ServletResponseConstants.SC_FILE_EXTENSION_EXCEPTION
+					).put(
+						"message", ".csv"
+					));
+			}
+			else {
+				throw portalException;
+			}
+
+			return jsonObject;
+		}
+
+		@Override
+		public JSONObject onSuccess(
+				UploadPortletRequest uploadPortletRequest, FileEntry fileEntry)
+			throws PortalException {
+
+			return _itemSelectorUploadResponseHandler.onSuccess(
+				uploadPortletRequest, fileEntry);
+		}
+
+	}
+
 	@Override
 	protected void doProcessAction(
 			ActionRequest actionRequest, ActionResponse actionResponse)
@@ -40,11 +174,31 @@ public class UploadCSVFileEntryMVCActionCommand extends BaseMVCActionCommand {
 			actionRequest, actionResponse);
 	}
 
-	@Reference
-	private CSVUploadFileEntryHandler _csvUploadFileEntryHandler;
+	private static final String _PARAMETER_NAME = "imageSelectorFileName";
+
+	private static final String _TEMP_FOLDER_NAME =
+		CSVUploadFileEntryHandler.class.getName();
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		UploadCSVFileEntryMVCActionCommand.class);
+
+	private final CSVUploadFileEntryHandler _csvUploadFileEntryHandler =
+		new CSVUploadFileEntryHandler();
+	private final CSVUploadResponseHandler _csvUploadResponseHandler =
+		new CSVUploadResponseHandler();
 
 	@Reference
-	private CSVUploadResponseHandler _csvUploadResponseHandler;
+	private DLValidator _dlValidator;
+
+	@Reference
+	private File _file;
+
+	@Reference
+	private ItemSelectorUploadResponseHandler
+		_itemSelectorUploadResponseHandler;
+
+	@Reference
+	private UniqueFileNameProvider _uniqueFileNameProvider;
 
 	@Reference
 	private UploadHandler _uploadHandler;
