--- modules/apps/commerce/commerce-order-content-web/src/main/java/com/liferay/commerce/order/content/web/internal/upload/CSVUploadResponseHandler.java
@@ -1,72 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.commerce.order.content.web.internal.upload;
-
-import com.liferay.document.library.kernel.exception.FileExtensionException;
-import com.liferay.item.selector.ItemSelectorUploadResponseHandler;
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.json.JSONFactory;
-import com.liferay.portal.kernel.json.JSONObject;
-import com.liferay.portal.kernel.repository.model.FileEntry;
-import com.liferay.portal.kernel.servlet.ServletResponseConstants;
-import com.liferay.portal.kernel.upload.UploadPortletRequest;
-import com.liferay.upload.UploadResponseHandler;
-
-import javax.portlet.PortletRequest;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Alessio Antonio Rendina
- */
-@Component(service = CSVUploadResponseHandler.class)
-public class CSVUploadResponseHandler implements UploadResponseHandler {
-
-	@Override
-	public JSONObject onFailure(
-			PortletRequest portletRequest, PortalException portalException)
-		throws PortalException {
-
-		JSONObject jsonObject = _itemSelectorUploadResponseHandler.onFailure(
-			portletRequest, portalException);
-
-		if (portalException instanceof FileExtensionException) {
-			JSONObject errorJSONObject = _jsonFactory.createJSONObject();
-
-			errorJSONObject.put(
-				"errorType",
-				ServletResponseConstants.SC_FILE_EXTENSION_EXCEPTION
-			).put(
-				"message", ".csv"
-			);
-
-			jsonObject.put("error", errorJSONObject);
-		}
-		else {
-			throw portalException;
-		}
-
-		return jsonObject;
-	}
-
-	@Override
-	public JSONObject onSuccess(
-			UploadPortletRequest uploadPortletRequest, FileEntry fileEntry)
-		throws PortalException {
-
-		return _itemSelectorUploadResponseHandler.onSuccess(
-			uploadPortletRequest, fileEntry);
-	}
-
-	@Reference
-	private ItemSelectorUploadResponseHandler
-		_itemSelectorUploadResponseHandler;
-
-	@Reference
-	private JSONFactory _jsonFactory;
-
-}
